test: move semantics layout tests under test/semantics/ (#1080)
Refs #1080

Made-with: Cursor
