+++ test/README.md
@@ -63,17 +63,17 @@ Representative files:
 
 | Area                                       | Start with                                                                                                                                                                                                                       | Notes                                                                                     |
 | ------------------------------------------ | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------- |
-| CLI behavior and artifact selection        | `cli/cli_contract_matrix.test.ts`, `cli/cli_failure_contract_matrix.test.ts`, `cli/cli_artifacts.test.ts`, `cli/cli_zax_smoke.test.ts`                                                                                         | Use `test/helpers/cli.ts` for end-to-end CLI execution.                                   |
+| CLI behavior and artifact selection        | `cli/cli_contract_matrix.test.ts`, `cli/cli_failure_contract_matrix.test.ts`, `cli/cli_artifacts.test.ts`, `cli/cli_zax_smoke.test.ts`                                                                                           | Use `test/helpers/cli.ts` for end-to-end CLI execution.                                   |
 | Parser dispatch and recovery               | `pr468_parser_dispatch_integration.test.ts`, `pr227_parser_toplevel_malformed_spans.test.ts`, `pr238_parser_malformed_decl_header_spans_matrix.test.ts`                                                                          | Use helper-level `pr476_parse_*.test.ts` files for isolated parser seams.                 |
 | Grammar and token tables                   | `pr762_grammar_data_conformance.test.ts`, `pr808_grammar_drift.test.ts`, `pr250_parser_instruction_head_casing.test.ts`, `pr252_parser_register_token_canonicalization.test.ts`                                                  | Good home for reserved-word and canonicalization changes.                                 |
-| Semantics and layout                       | `semantics_layout.test.ts`, `semantics_layout_extra.test.ts`, `pr285_alias_init_parser_semantics_matrix.test.ts`, `pr980_local_alias_legality.test.ts`                                                                           | Use these when type size, offsets, alias legality, or compile-time rules change.          |
+| Semantics and layout                       | `semantics/semantics_layout.test.ts`, `semantics/semantics_layout_extra.test.ts`, `pr285_alias_init_parser_semantics_matrix.test.ts`, `pr980_local_alias_legality.test.ts`                                                       | Use these when type size, offsets, alias legality, or compile-time rules change.          |
 | Lowering frame and control-flow invariants | `pr102_lowering_frame_invariants.test.ts`, `pr103_lowering_mixed_return_paths.test.ts`, `pr555_function_sp_state_integration.test.ts`, `pr848_break_continue_integration.test.ts`                                                | Covers epilogue cleanup, SP tracking, and structured control lowering.                    |
 | Lowering helper seams                      | `pr509_*.test.ts`, `pr510_*.test.ts`, `pr528_emission_core_helpers.test.ts`, `pr529_fixup_emission_helpers.test.ts`, `pr530_asm_utils_helpers.test.ts`, `pr531_value_materialization_helpers.test.ts`                            | Prefer these before adding another broad compile test.                                    |
 | Named sections and placement               | `pr572_named_sections_parser.test.ts`, `pr582_section_contribution_sinks.test.ts`, `pr583_section_placement_helpers.test.ts`, `pr584_named_section_fixups_integration.test.ts`, `pr585_named_section_layout_integration.test.ts` | Use when section routing, anchors, or fixups change.                                      |
 | Assignment and storage legality            | `pr862_assignment_parser.test.ts`, `pr863_assignment_lowering.test.ts`, `pr869_assignment_reg8_*`, `pr875_assignment_ixiy_integration.test.ts`, `pr895_assignment_acceptance.test.ts`                                            | Parser, lowering, and acceptance coverage are already split.                              |
 | Encoder behavior                           | `pr24_isa_core.test.ts`, `backend/pr477_encode_*.test.ts`, `pr203_ld_diag_matrix.test.ts`, `pr240_isa_register_target_diag_matrix.test.ts`                                                                                       | Prefer direct encoder tests when lowering is not involved.                                |
 | Examples, smoke, and determinism           | `examples_compile.test.ts`, `smoke.test.ts`, `smoke_language_tour_compile.test.ts`, `determinism_artifacts.test.ts`                                                                                                              | Use for broad regressions and checked-in examples.                                        |
-| Corpus and external backend compatibility  | `pr680_asm_golden_contract.test.ts`, `cli/pr990_asm80_emitter_validation.test.ts`, `pr991_asm80_comment_preservation.test.ts`                                                                                                  | Only use these when emitted artifact text or external-tool compatibility is the contract. |
+| Corpus and external backend compatibility  | `pr680_asm_golden_contract.test.ts`, `cli/pr990_asm80_emitter_validation.test.ts`, `pr991_asm80_comment_preservation.test.ts`                                                                                                    | Only use these when emitted artifact text or external-tool compatibility is the contract. |
 | Policy and infrastructure                  | `ci_change_classifier.test.ts`, `pr472_source_file_size_guard.test.ts`, `pr241_d8m_contract_hardening.test.ts`                                                                                                                   | For repo policy, CI classification, and artifact contract checks.                         |
 
 ## Where to put new tests
